requirements.txt
pytest~=7.1.2
pytest-cov
pytest-flake8
pytest-spec

py~=1.11.0
colorama~=0.4.5
pip~=21.3.1
attrs~=21.4.0
toml~=0.10.2
wheel~=0.37.1
flake8~=4.0.1
pyflakes~=2.4.0
pluggy~=1.0.0
iniconfig~=1.1.1
atomicwrites~=1.4.1
coverage~=6.4.2
pyparsing~=3.0.9
setuptools~=60.2.0
packaging~=21.3
requests~=2.28.1
certifi~=2022.6.15
six~=1.16.0
idna~=3.3
urllib3~=1.26.10
Faker~=13.15.0
PyYAML~=6.0
python-dateutil~=2.8.2